+++ src/common/messages.h
@@ -18,7 +18,6 @@ struct bilingual_str;
 
 enum class FeeEstimateMode;
 enum class FeeReason;
-enum class BlockPolicyEstimateReason;
 namespace node {
 enum class TransactionError;
 } // namespace node
@@ -27,7 +26,6 @@ namespace common {
 enum class PSBTError;
 bool FeeModeFromString(std::string_view mode_string, FeeEstimateMode& fee_estimate_mode);
 std::string StringForFeeReason(FeeReason reason);
-std::string StringForBlockPolicyEstimateReason(BlockPolicyEstimateReason reason);
 std::string FeeModes(const std::string& delimiter);
 std::string FeeModeInfo(std::pair<std::string, FeeEstimateMode>& mode);
 std::string FeeModesDetail(std::string default_info);
